Extract private methods from Api into HttpService
